error messages for tasks_watchers project_users users_teams updated, mail on task update added

+++ app/controllers/projects_controller.rb
@@ -41,11 +41,7 @@ class ProjectsController < ApplicationController
     if @project.save
       flash[:notice] = t 'project.created'
     else
-      errors = ''
-      @project.errors.full_messages.each do |msg|
-        errors = errors + msg + ', '
-      end
-      flash[:error] = errors
+      flash[:error] = @project.errors.full_messages
       success = false
     end
     respond_to do |format|
@@ -63,6 +59,7 @@ class ProjectsController < ApplicationController
       flash[:notice] = t 'project.updated'
     else
       success = false
+      flash[:error] = @project.errors.full_messages
     end
     respond_to do |format|
       if success
@@ -74,11 +71,24 @@ class ProjectsController < ApplicationController
   end
 
   def destroy
-    if @project.destroy
-      flash[:notice] = t 'project.destroyed'
+    if @project.can_destroy?
+      if @project.destroy
+        flash[:notice] = t('project.destroyed')
+      else
+        flash[:error] = @project.errors.full_messages
+      end
+    else
+      flash[:error] = @project.errors.full_messages
     end
+
     respond_to do |format|
-      format.html { redirect_to @project }
+      format.html do
+        if flash[:error].blank?
+          redirect_to projects_path
+        else
+          redirect_to @project
+        end
+      end
     end
   end
